requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
open_clip_torch>=2.24.0
pillow>=9.0.0
pandas>=1.5.0
scikit-learn>=1.2.0
tqdm>=4.65.0
numpy>=1.24.0
matplotlib>=3.7.0
